Audio engine - playback speed

+++ EngineSoundTest/audio.h
@@ -10,7 +10,7 @@ struct Audio {
 	std::vector<float> samples = {};
     std::vector<Grain> grains = {};
     int sampleCount = 0;
-    int cursor = 0;
+    float cursor = 0.0f;
     int currentGrainId = 0;
 };
